requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
sqlalchemy==2.0.27
pydantic==2.6.1
python-dotenv==1.0.1
google-generativeai==0.3.2
numpy==1.26.4
pandas==2.2.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
streamlit==1.31.1 